pyproject.toml
[tool.poetry]
name = "applicationName"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
infi-systray = "^0.1.12"
pypiwin32 = "^223"
pillow = "^9.5.0"
pystray = "^0.19.4"
python-vlc = "^3.0.18122"

[tool.poetry.dev-dependencies]


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pyinstaller = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
